move search headless and analytics providers into component
